requirements.txt
annotated-types==0.7.0
anyio==4.11.0
backports.asyncio.runner==1.2.0
boto3==1.40.54
botocore==1.40.54
certifi==2025.10.5
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
s3transfer==0.14.0
sentry-sdk==2.42.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tomli==2.3.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
